in progress

+++ src/interval.zig
@@ -1,26 +1,26 @@
 const std = @import("std");
 
 pub const Interval = struct {
-    min: f32,
-    max: f32,
+    min: f64,
+    max: f64,
 
-    pub fn init(min: f32, max: f32) Interval {
+    pub fn init(min: f64, max: f64) Interval {
         return .{ .min = min, .max = max };
     }
 
-    pub fn size(self: Interval) f32 {
+    pub fn size(self: Interval) f64 {
         return self.max - self.min;
     }
 
-    pub fn contains(self: Interval, value: f32) bool {
+    pub fn contains(self: Interval, value: f64) bool {
         return self.min <= value and value <= self.max;
     }
 
-    pub fn surrounds(self: Interval, value: f32) bool {
+    pub fn surrounds(self: Interval, value: f64) bool {
         return self.min < value and value < self.max;
     }
 
-    pub fn clamp(self: Interval, value: f32) f32 {
+    pub fn clamp(self: Interval, value: f64) f64 {
         if (value < self.min) {
             return self.min;
         }
@@ -33,5 +33,5 @@ pub const Interval = struct {
     }
 };
 
-pub const empty = Interval.init(std.math.floatMax(f32), std.math.floatMin(f32));
-pub const universe = Interval.init(std.math.floatMin(f32), std.math.floatMax(f32));
+pub const empty = Interval.init(std.math.floatMax(f64), std.math.floatMin(f64));
+pub const universe = Interval.init(std.math.floatMin(f64), std.math.floatMax(f64));
